+++ limits.c
@@ -15,8 +15,8 @@
 
 #define READ_TITLE(ch) \
    ( GET_SEX(ch) == SEX_MALE ? \
-    titles[GET_CLASS(ch)-1][GET_LEVEL(ch)].title_m : \
-    titles[GET_CLASS(ch)-1][GET_LEVEL(ch)].title_f )
+    titles[(int)GET_CLASS(ch)-1][(int)GET_LEVEL(ch)].title_m : \
+    titles[(int)GET_CLASS(ch)-1][(int)GET_LEVEL(ch)].title_f )
 
 
 extern struct char_data *character_list;
